+++ .eslintrc
@@ -2,7 +2,7 @@ root: true
 
 parserOptions:
   ecmaVersion: 2018
-  sourceType: "script"
+  sourceType: "module"
 
 env:
   es2017: true
